Finish login and signup page

+++ pybo/views/question_views.py
@@ -6,17 +6,17 @@ from django.utils import timezone
 from pybo.forms import QuestionForm
 from pybo.models import Question
 
-from django.contrib.auth.models import User #관리자 권한 부여
-user = User.objects.get(username='admin') # 관리자 권한 부여
+# from django.contrib.auth.models import User #관리자 권한 부여
+# user = User.objects.get(username='admin') # 관리자 권한 부여
 
 
-# @login_required(login_url='common:login')
+@login_required(login_url='common:login')
 def question_create(request):
     if request.method == 'POST':
         form = QuestionForm(request.POST)
         if form.is_valid():
             question = form.save(commit=False)
-            question.author = user  # author 속성에 로그인 계정 저장
+            question.author = request.user  # author 속성에 로그인 계정 저장
             question.create_date = timezone.now()
           
             question.type = "free"  # type 속성 추가
@@ -28,12 +28,13 @@ def question_create(request):
     context = {'form': form}
     return render(request, 'pybo/question_form.html', context)
 
+@login_required(login_url='common:login')
 def question_create_2(request):
     if request.method == 'POST':
         form = QuestionForm(request.POST)
         if form.is_valid():
             question = form.save(commit=False)
-            question.author = user  # author 속성에 로그인 계정 저장
+            question.author = request.user  # author 속성에 로그인 계정 저장
             question.create_date = timezone.now()
           
             question.type = "info"  # type 속성 추가
@@ -45,12 +46,13 @@ def question_create_2(request):
     context = {'form': form}
     return render(request, 'pybo/question_form.html', context)
 
+@login_required(login_url='common:login')
 def question_create_3(request):
     if request.method == 'POST':
         form = QuestionForm(request.POST)
         if form.is_valid():
             question = form.save(commit=False)
-            question.author = user  # author 속성에 로그인 계정 저장
+            question.author = request.user # author 속성에 로그인 계정 저장
             question.create_date = timezone.now()
           
             question.type = "pr"  # type 속성 추가
@@ -62,12 +64,13 @@ def question_create_3(request):
     context = {'form': form}
     return render(request, 'pybo/question_form.html', context)
 
+@login_required(login_url='common:login')
 def question_create_4(request):
     if request.method == 'POST':
         form = QuestionForm(request.POST)
         if form.is_valid():
             question = form.save(commit=False)
-            question.author = user  # author 속성에 로그인 계정 저장
+            question.author = request.user  # author 속성에 로그인 계정 저장
             question.create_date = timezone.now()
           
             question.type = "sale"  # type 속성 추가
@@ -80,12 +83,12 @@ def question_create_4(request):
     return render(request, 'pybo/question_form.html', context)
 
 
-# @login_required(login_url='common:login')
+@login_required(login_url='common:login')
 def question_modify(request, question_id):
     question = get_object_or_404(Question, pk=question_id)
-    # if request.user != question.author:
-    #     messages.error(request, '수정권한이 없습니다')
-    #     return redirect('pybo:detail', question_id=question.id)
+    if request.user != question.author:
+        messages.error(request, '수정권한이 없습니다')
+        return redirect('pybo:detail', question_id=question.id)
     if request.method == "POST":
         form = QuestionForm(request.POST, instance=question)
         if form.is_valid():
@@ -99,21 +102,21 @@ def question_modify(request, question_id):
     return render(request, 'pybo/question_form.html', context)
 
 
-# @login_required(login_url='common:login')
+@login_required(login_url='common:login')
 def question_delete(request, question_id):
     question = get_object_or_404(Question, pk=question_id)
-    # if request.user != question.author:
-    #     messages.error(request, '삭제권한이 없습니다')
-    #     return redirect('pybo:detail', question_id=question.id)
+    if request.user != question.author:
+        messages.error(request, '삭제권한이 없습니다')
+        return redirect('pybo:detail', question_id=question.id)
     question.delete()
     return redirect('pybo:index')
 
 
-# @login_required(login_url='common:login')
+@login_required(login_url='common:login')
 def question_vote(request, question_id):
     question = get_object_or_404(Question, pk=question_id)
-    # if request.user == question.author:
-    #     messages.error(request, '본인이 작성한 글은 추천할수 없습니다')
-    # else:
-    question.voter.add(user)
+    if request.user == question.author:
+        messages.error(request, '본인이 작성한 글은 추천할수 없습니다')
+    else:
+      question.voter.add(request.user)
     return redirect('pybo:detail', question_id=question.id)
